refactor

+++ src/smtrat-cad/utils/MISGeneration.h
@@ -132,7 +132,7 @@ namespace cad {
 			[](const ConstraintT& c){
 				constexpr double constant_weight   = 5.0;
 				constexpr double complexity_weight = 0.1;
-				return constant_weight + complexity_weight * complexity(c);
+				return constant_weight + complexity_weight * (double)complexity(c);
 			}
 		);
